+++ src/Server/Persistence/Database/Sqlite/Queries/SqliteRetrieveEntityQuery.cs
@@ -30,16 +30,16 @@ public sealed class SqliteRetrieveEntityQuery : IRetrieveEntityQuery
     ///     SQL query to execute.
     /// </summary>
     private const string query =
-        @"WITH RECURSIVE EntityTree(id, x, y, z, material, materialModifier, playerCharacter, name, account,parent)
+        @"WITH RECURSIVE EntityTree(id, x, y, z, material, materialModifier, playerCharacter, name, account, parent, drawable)
 	        AS (
-		        SELECT id, x, y, z, material, materialModifier, playerCharacter, name, account, parent
+		        SELECT id, x, y, z, material, materialModifier, playerCharacter, name, account, parent, drawable
 			        FROM EntityWithComponents WHERE id = @Id
 		        UNION ALL
-            		SELECT ec.id, ec.x, ec.y, ec.z, ec.material, ec.materialModifier, ec.playerCharacter, ec.name, ec.account, ec.parent
+            		SELECT ec.id, ec.x, ec.y, ec.z, ec.material, ec.materialModifier, ec.playerCharacter, ec.name, ec.account, ec.parent, ec.drawable
 			        FROM EntityWithComponents ec, EntityTree et
         			WHERE ec.parent = et.id
 	        )
-            SELECT id, x, y, z, material, materialModifier, playerCharacter, name, account, parent FROM EntityTree";
+            SELECT id, x, y, z, material, materialModifier, playerCharacter, name, account, parent, drawable FROM EntityTree";
 
     private readonly SqliteConnection dbConnection;